# By Anton Chervyakovskiy (8) and Sergey Berzhinsky (7)
# Via Anton Chervyakovsky (3) and Anton Chervyakovskiy (2)
* 'develop' of https://github.com/nikolay-kvetko/DataManager:
  Bug: bug fixed add new date field, implement create/edit dateField on instance page
  Bug: fixed add ImageUrl with incorrect url
  Task: AR009604-179 Introduce GPS coordinates field on Add/Edit form for Entity Instances management - bug fixed image and gps fields with required param
  Task: AR009604-179 Introduce GPS coordinates field on Add/Edit form for Entity Instances management - create GPSValue class, implement create/edit instance with GPSField on modal window, implement display GPS value on instance list view
  Task AR009604-186: Implement Company creation form - implemented Company creation form with possibility to download Company logo
  Task AR009604-182: Introduce Company into DataManager Application - users of any Company have access only to Entity Instances (watch,delete, update) that belong to their Company
  Task AR009604-182: Introduce Company into DataManager Application - users of any Company have access only to Fields (watch,delete, update) that belong to their Company
  Task AR009604-182: Introduce Company into DataManager Application - users of any Company have access only to Entity Schemas (watch,delete, update) that belong to their Company
  Task AR009604-186: Implement Company creation form - added beta Company creation form; Task AR009604-200: Allow access to Data Manager application only to logged in users - user with role Admin can't finish registration without creation own company
  Task AR009604-182: Introduce Company into DataManager Application - added new database table "Company"; added relations between "Company" and  "Entity Schema", "Company" and  "User"; added CRUD operations for Company and cover them by necessary tests
  Task AR009604-200: Allow access to Data Manager application only to logged in users - added security layer
  Task: AR009604-177 Implement Create/Edit Form on Modal Dialog for GPS coordinates field - create GPSField, implement modal window for create/edit GPSField in dataStructure
  Task: add possibility create/edit imageField for entity instances management
  Task AR009604-188: Introduce User into Data Manager application - added User and Role classes with the necessary relationships; added some CRUD operations and covered them by tests
  Bug: bug fixed add new date field, implement create/edit dateField on instance page

+++ src/main/java/com/intetics/controller/EntitySchemaController.java
@@ -1,10 +1,15 @@
 package com.intetics.controller;
 
+import com.intetics.bean.Company;
 import com.intetics.bean.EntitySchema;
+import com.intetics.bean.User;
 import com.intetics.dao.EntitySchemaDao;
+import com.intetics.dao.UserDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.util.Assert;
@@ -13,8 +18,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.annotation.Nonnull;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
+import java.security.Principal;
 import java.util.Date;
 import java.util.List;
 
@@ -30,10 +34,14 @@ public class EntitySchemaController {
     @Autowired
     private EntitySchemaDao entitySchemaDao;
 
-    @RequestMapping(value = "/list")
-    public String getEntitySchemaList(ModelMap model) {
+    @Autowired
+    private UserDao userDao;
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+    @RequestMapping(value = "/list")
+    public String getEntitySchemaList(ModelMap model, Principal principal) {
+
+        User user = userDao.getUserByEmail(principal.getName());
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         LOGGER.trace("The list of entities has been requested");
@@ -42,13 +50,13 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/create")
-    public String createEntitySchema(ModelMap model) {
+    public String createEntitySchema(ModelMap model, Principal principal) {
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+        User user = userDao.getUserByEmail(principal.getName());
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         EntitySchema entitySchema = new EntitySchema();
-        entitySchema.setName("");
         model.addAttribute("EntitySchema", entitySchema);
 
         model.addAttribute("modalTitle", "Create Entity");
@@ -60,16 +68,14 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/edit/{entitySchemaId}", method = RequestMethod.GET)
-    public String startToEditEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model,
-                                          HttpServletRequest request) {
+    public String startToEditEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         model.addAttribute("EntitySchema", entitySchema);
-
-        HttpSession session = request.getSession();
-        session.setAttribute("entitySchemaCreateDate-"+entitySchemaId, entitySchema.getCreateDate());
-
         model.addAttribute("modalTitle", "Edit Entity");
         model.addAttribute("modalSaveButton", "Edit");
 
@@ -77,33 +83,45 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
-    public String saveEntitySchema(EntitySchema entitySchema, HttpServletRequest request) {
+    public String saveEntitySchema(EntitySchema entitySchema, Principal principal) {
 
         Date currentDate = new Date();
 
+        User user = userDao.getUserByEmail(principal.getName());
+
+        Company company = user.getCompany();
+        entitySchema.setCompany(company);
+
         if(entitySchema.getId() != null){
-            HttpSession session = request.getSession();
-            entitySchema.setCreateDate((Date) session.getAttribute("entitySchemaCreateDate-"+entitySchema.getId()));
+            if(verifyComplianceEntitySchemaAndCompany(entitySchema.getId()) == null)
+                return "error";
+
             entitySchema.setModifiedDate(currentDate);
             entitySchemaDao.saveOrUpdate(entitySchema);
             return "redirect:/entity/"+entitySchema.getId()+"/field/list";
         }
 
-        entitySchema.setCreateDate(currentDate);
         entitySchema.setModifiedDate(currentDate);
+        entitySchema.setCompany(company);
 
         entitySchemaDao.saveOrUpdate(entitySchema);
         return "redirect:/entity/list";
     }
 
     @RequestMapping(value = "/delete/{entitySchemaId}/confirm", method = RequestMethod.GET)
-    public String startDeleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model) {
+    public String startDeleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model,
+                                          Principal principal) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         model.addAttribute("EntitySchema", entitySchema);
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+        User user = userDao.getUserByEmail(principal.getName());
+
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         return "delete-entity";
@@ -113,9 +131,28 @@ public class EntitySchemaController {
     public String deleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         entitySchemaDao.delete(entitySchema);
 
         return "redirect:/entity/list";
+    }
+
+    private EntitySchema verifyComplianceEntitySchemaAndCompany(Long entitySchemaId) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = userDao.getUserByEmail(authentication.getName());
+
+        Company company = user.getCompany();
+
+        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+
+        if(company.getEntitySchemas().contains(entitySchema))
+            return entitySchema;
+
+        return null;
     }
 }
